+++ src/main/java/com/fastcompus/springpractice/logic/JavaSort.java
@@ -1,9 +1,12 @@
-package com.fastcompus.javaoop.logic;
+package com.fastcompus.springpractice.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+@Component
 public class JavaSort<T extends Comparable<T>> implements Sort<T> {
     public List<T> sort(List<T> list) {
         List<T> output = new ArrayList<>(list);
