+++ src/test/java/com/fastcompus/springpractice/MainTest.java
@@ -1,8 +1,6 @@
-package com.fastcompus.javaoop;
+package com.fastcompus.springpractice;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class MainTest {
