+++ src/test/java/com/fastcompus/springpractice/logic/JavaSortTest.java
@@ -1,4 +1,4 @@
-package com.fastcompus.javaoop.logic;
+package com.fastcompus.springpractice.logic;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,16 +7,15 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class BubbleSortTest {
-
-    @DisplayName("버블소트 - 리스트를 넣으면 정렬된 결과를 보여준다")
+class JavaSortTest {
+    @DisplayName("자바소트 - 리스트를 넣으면 정렬된 결과를 보여준다")
     @Test
     void given_List_WhenExecution_ThenReturnSortedList() {
         // Given
-        BubbleSort<Integer> bubbleSort = new BubbleSort<>();
+        JavaSort<Integer> javaSort = new JavaSort<>();
 
         // When
-        List<Integer> actual = bubbleSort.sort(List.of(3, 2, 4, 5, 1));
+        List<Integer> actual = javaSort.sort(List.of(3, 2, 4, 5, 1));
 
         // Then
         assertEquals(List.of(1,2,3,4,5), actual);
